+++ otter/assign/r_adapter/solutions.py
@@ -1,24 +1,24 @@
 """Solution removal for Otter Assign for R notebooks"""
 
-import copy
 import re
 
-from ..constants import MD_RESPONSE_CELL_SOURCE
-from ..utils import get_source
+
+BLOCK_PROMPT = "# YOUR CODE HERE"
+
 
 solution_assignment_regex = re.compile(r"(\s*[\w. $()]*(=|<-))(.*) ?# ?SOLUTION")
 def solution_assignment_sub(match):
     """
-    Substitutes the first matching group  with ` ...`
+    Substitutes the first matching group  with `` NULL # YOUR CODE HERE``
     """
     prefix = match.group(1)
-    return prefix + ' ...'
+    return prefix + ' NULL # YOUR CODE HERE'
 
 
 solution_line_regex = re.compile(r"(\s*).* ?# ?SOLUTION")
 def solution_line_sub(match):
     """
-    Substitutes the first matching group  with `# YOUR CODE HERE`
+    Substitutes the first matching group  with ``# YOUR CODE HERE``
     """
     prefix = match.group(1)
     return prefix + '# YOUR CODE HERE'