+++ otter/assign/v0/rmarkdown_adapter/tests.py
@@ -1,5 +1,7 @@
 import re
 
+from dataclasses import replace
+
 from .utils import Cell
 
 from ..constants import TEST_REGEX
@@ -9,7 +11,7 @@ from ..utils import get_source
 def is_test_cell(cell):
     """
     Returns whether the current cell is a test cell
-    
+
     Args:
         cell (``otter.assign.rmarkdown_adapter.utils.Cell``): an Rmd file cell
 
@@ -26,7 +28,7 @@ def gen_test_cell(question, tests, tests_dict, assignment):
     Parses a list of test named tuples and creates a single test file. Adds this test file as a value
     to ``tests_dict`` with a key corresponding to the test's name, taken from ``question``. Returns
     a code cell that runs the check on this test.
-    
+
     Args:
         question (``dict``): question metadata
         tests (``list`` of ``Test``): tests to be written
@@ -49,7 +51,7 @@ def gen_test_cell(question, tests, tests_dict, assignment):
         f"Points for question {question['name']} could not be parsed:\n{points}"
 
     # update point values
-    tests = [tc._replace(points=p) for tc, p in zip(tests, points)]
+    tests = [replace(tc, points=p) for tc, p in zip(tests, points)]
     test = gen_suite(question['name'], tests, points)
 
     tests_dict[question['name']] = test