+++ otter/execute/execute_log.py
@@ -31,7 +31,7 @@ def execute_log(nb, log, check_results_list_name="check_results_secret", initial
         test_dir (``str``, optional): path to directory of tests in grading environment
         variables (``dict``, optional): map of variable names -> type string to check type of deserialized
             object to prevent arbitrary code from being put into the environment
-    
+
     Results:
         ``dict``: global environment resulting from executing all code of the input notebook
     """
@@ -44,7 +44,7 @@ def execute_log(nb, log, check_results_list_name="check_results_secret", initial
         source = f"import otter\ngrader = otter.Notebook(\"{test_dir}\")\n"
     else:
         source = f"import otter\ngrader = otter.Notebook()\n"
-    
+
     if cwd:
         source +=  f"import sys\nsys.path.append(\"{cwd}\")\n"
 
@@ -75,7 +75,7 @@ def execute_log(nb, log, check_results_list_name="check_results_secret", initial
                 # TODO: this is a horrible implementation, make it better! maybe use an AST and
                 # astunparse?
                 cell_source_lines = [re.sub(r"^\s+", "", l) for l in cell_source_lines if "import" in l]                                
-                
+
                 for line in cell_source_lines:
                     try:
                         exec(line, global_env)
@@ -100,5 +100,5 @@ def execute_log(nb, log, check_results_list_name="check_results_secret", initial
             logged_questions.append(entry.question)
 
     print("Questions executed from log: {}".format(", ".join(logged_questions)))
-    
+
     return global_env