+++ otter/export/exporters/utils.py
@@ -1,6 +1,4 @@
-"""
-Utils for Otter Export exporterss
-"""
+"""Utilities for Otter Export exporters"""
 
 import re
 import copy
@@ -8,12 +6,14 @@ import nbformat
 
 from ...utils import get_source
 
+
 BEGIN_QUESTION_REGEX = r"<!--\s*BEGIN QUESTION\s*-->"
 END_QUESTION_REGEX = r"<!--\s*END QUESTION\s*-->"
 NEW_PAGE_REGEX = r"<!--\s*NEW PAGE\s*-->"
 NEW_PAGE_MARKER = "#newpage"
 NEW_PAGE_CELL_SOURCE = f"<!-- {NEW_PAGE_MARKER} -->"
 NEW_PAGE_CLASS_NAME = "otter-page-break-after"
+
 
 def has_begin(line):
     """
@@ -25,14 +25,15 @@ def has_begin(line):
     .. code-block:: html
 
         <!-- BEGIN QUESTION -->
-    
+
     Args:
         line (``str``): the line to search
-    
+
     Returns:
         ``bool``: whether the line contains a substring matching the begin question regex
     """
     return bool(re.search(BEGIN_QUESTION_REGEX, line, flags=re.IGNORECASE))
+
 
 def has_end(line):
     """
@@ -44,14 +45,15 @@ def has_end(line):
     .. code-block:: html
 
         <!-- END QUESTION -->
-    
+
     Args:
         line (``str``): the line to search
-    
+
     Returns:
         ``bool``: whether the line contains a substring matching the end question regex
     """
     return bool(re.search(END_QUESTION_REGEX, line, flags=re.IGNORECASE))
+
 
 def sub_end_for_new_page(line):
     """
@@ -71,6 +73,7 @@ def sub_end_for_new_page(line):
         ``str``: the line with the end question match substituted for the newpage comment
     """
     return re.sub(END_QUESTION_REGEX, NEW_PAGE_CELL_SOURCE, line)
+
 
 def notebook_pdf_generator(nb):
     """
@@ -103,12 +106,12 @@ def notebook_pdf_generator(nb):
             all_cells.append(subnb_cells)
             subnb_cells = []
             subnb_cells.append(c2)
-            
+
         else:
             subnb_cells.append(cell)
-    
+
     all_cells.append(subnb_cells)
-    
+
     for subnb_cells in all_cells:
         dummy_nb.cells = subnb_cells
         yield dummy_nb