+++ otter/grade/__init__.py
@@ -3,6 +3,8 @@
 import os
 import pandas as pd
 import re
+import shutil
+import tempfile
 
 from .containers import launch_grade
 from .utils import merge_csv, prune_images
@@ -46,6 +48,20 @@ def main(*, path="./", output_dir="./", autograder="./autograder.zip", container
         prune_images(force=force)
         return
 
+    # if path leads to single file this indicates
+    # the case and changes path to the directory
+    # as well as updates the ext argument
+    single_file = False
+    temp_file_path = ""
+    if os.path.isfile(path):
+        single_file = True
+        ext = os.path.splitext(path)[1][1:]  # remove the period from extension
+        file = os.path.split(path)[1]
+        temp_dir = tempfile.mkdtemp(prefix="otter_")
+        temp_file_path = f"{temp_dir}/{file}"
+        shutil.copy(path, temp_file_path)
+        path = temp_dir
+
     # check file paths
     assert_path_exists([
         (path, True),
@@ -81,3 +97,6 @@ def main(*, path="./", output_dir="./", autograder="./autograder.zip", container
 
     # write to CSV file
     output_df.to_csv(os.path.join(output_dir, "final_grades.csv"), index=False)
+    if single_file:
+        shutil.rmtree(temp_dir)
+        return output_df["percent_correct"][0]