+++ test/test-assign/jupyterlite-correct/student/tests/q1.py
@@ -1,0 +1,21 @@
+OK_FORMAT = True
+
+test = {   'name': 'q1',
+    'points': None,
+    'suites': [   {   'cases': [   {   'code': '>>> isinstance(x, int)\nTrue',
+                                       'failure_message': 'This is not an int.',
+                                       'hidden': False,
+                                       'locked': False,
+                                       'points': 2,
+                                       'success_message': 'Congrats you passed this test case!\\'},
+                                   {'code': '>>> None\n', 'hidden': False, 'locked': False, 'points': 3, 'success_message': 'Congrats, this passed!'},
+                                   {   'code': '>>> 0 < x < 100\nTrue',
+                                       'failure_message': 'This should have passed.',
+                                       'hidden': False,
+                                       'locked': False,
+                                       'points': 1,
+                                       'success_message': 'Congrats your x value is in the correct range!'}],
+                      'scored': True,
+                      'setup': '',
+                      'teardown': '',
+                      'type': 'doctest'}]}