+++ test/test-assign/rmd-correct/student/tests/q3.R
@@ -6,9 +6,7 @@ test = list(
       name = NA,
       points = 0.6666666666666666,
       code = {
-        test_that("q3a", {
-            expect_equal(nine, 9)
-        })
+        testthat::expect_equal(nine, 9)
       }
     ),
     ottr::TestCase$new(
@@ -16,11 +14,8 @@ test = list(
       name = NA,
       points = 0.6666666666666666,
       code = {
-        test_that("q3b", {
-            expect_equal(square(16), 256)
-        })
+        testthat::expect_equal(square(16), 256)
       }
     )
-
   )
 )